keep more functions in default_stage, so that we can change it in control.go.
